pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tstriage"
dynamic = ["version"]
description = "MEPG TS Triage Runner"
readme = "README.md"
authors = [
    {name = "poke30744", email = "poke30744@gmail.com"}
]
license = "MIT"
requires-python = ">=3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyYAML",
    "psutil",
    "ffmpeg-python",
    "pysubs2",
    "tqdm",
]

[project.scripts]
tstriage = "tstriage.runner:main"

[project.urls]
Homepage = "https://github.com/poke30744/tstriage/blob/main/README.md"

[tool.setuptools]
packages = ["tstriage"]
include-package-data = true

[tool.setuptools.package-data]
tstriage = ["event.yml"]

[tool.setuptools.dynamic]
version = {attr = "tstriage.__version__"}

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]

[tool.uv]
required-version = ">=0.9.0"